🐛 fix(类型): 全量修复 INTEGER 列的 i64→i32 类型不匹配
- category.rs: sort_order, level
- unit.rs: decimal_places, sort_order
- material.rs: decimal_places (UnitOption)
- bom.rs: sort_order
- replenishment.rs: analysis_days, lead_days, safety_days（结构体+函数签名+局部变量）

+++ src-tauri/src/commands/bom.rs
@@ -57,7 +57,7 @@ pub struct BomItemDetail {
     pub substitute_id: Option<i64>,
     pub substitute_name: Option<String>,
     pub remark: Option<String>,
-    pub sort_order: i64,
+    pub sort_order: i32,
 }
 
 /// BOM 详情（头 + 明细）
@@ -119,7 +119,7 @@ pub struct SaveBomItemParams {
     pub is_key_part: bool,
     pub substitute_id: Option<i64>,
     pub remark: Option<String>,
-    pub sort_order: i64,
+    pub sort_order: i32,
 }
 
 /// 物料反查结果
